cast to ULL

+++ fflas-ffpack/fflas/fflas_bounds.inl
@@ -5,20 +5,20 @@
  *
  * Written by Clement Pernet <Clement.Pernet@imag.fr>
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library FFLAS-FFPACK.
- * 
+ *
  * FFLAS-FFPACK is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -129,6 +129,7 @@ namespace FFLAS {
 				if (d < 2)
 					return 1;
 				kmax = floor (d * double(1ULL << w));
+				// if (kmax  <= 1) return 1;
 			} else {
 
 				double c = computeFactorClassic(F);
@@ -143,12 +144,11 @@ namespace FFLAS {
 					}
 				}
 				kmax = floor ( (double (double(1ULL << mantissa) - cplt)) / (c*c));
-				if (kmax  <= 1)
-					return 1;
+				if (kmax  <= 1) return 1;
 			}
 
 			//kmax--; // we computed a strict upper bound
-			return  (size_t) MIN ((uint64_t)kmax, 1ULL << 31);
+			return  (size_t) MIN ((unsigned long long)kmax, 1ULL << 31);
 		}
 
 		/** @internal
